car_finance_scrapy/spiders/requirements.txt
alabaster==0.7.10
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
awscli==1.11.126
Babel==2.4.0
beautifulsoup4==4.6.0
botocore==1.5.89
bs4==0.0.1
cairocffi==0.8.0
certifi==2017.4.17
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
configparser==3.7.4
constantly==15.1.0
cssselect==1.0.3
Cython==0.28.1
decorator==4.2.1
Django==2.1b1
docker-py==1.10.6
docker-pycreds==0.3.0
docutils==0.13.1
Flask==1.0.2
html-parser==0.2
HTMLParser==0.0.2
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
lxml==4.2.0
MarkupSafe==1.0
msgpack-python==0.5.6
mysql-connector-python==8.0.12
numpy==1.14.2
pandas==0.22.0
parse==1.12.0
parsel==1.4.0
parso==0.1.1
peewee==3.5.2
pexpect==4.4.0
pickleshare==0.7.4
ply==3.11
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyMySQL==0.9.3
pypi==2.1
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
queuelib==1.5.0
requests==2.19.1
retrying==1.3.3
rsa==3.4.2
s3transfer==0.1.10
scrapinghub==2.0.3
Scrapy==1.5.0
selenium==3.141.0
service-identity==17.0.0
shub==2.9.0
simplegeneric==0.8.1
six==1.10.0
toml==0.10.0
tqdm==4.25.0
traitlets==4.3.2
Twisted==17.9.0
unicodecsv==0.14.1
urllib3==1.22
w3lib==1.19.0
wcwidth==0.1.7
websocket-client==0.52.0
Werkzeug==0.14.1
xmltodict==0.11.0
zope.interface==4.4.3
